+++ src/config/parse.rs
@@ -3,8 +3,11 @@ use clap::{value_t, values_t, ArgMatches};
 use coveralls_api::CiService;
 use log::error;
 use regex::Regex;
+use serde::de::{self, Deserializer};
 use std::env;
+use std::fmt;
 use std::path::PathBuf;
+use std::str::FromStr;
 use std::time::Duration;
 
 pub(super) fn get_list(args: &ArgMatches, key: &str) -> Vec<String> {
@@ -44,6 +47,12 @@ pub(super) fn get_manifest(args: &ArgMatches) -> PathBuf {
         manifest.push(path);
     }
 
+    manifest.push("Cargo.toml");
+    manifest.canonicalize().unwrap_or(manifest)
+}
+
+pub(super) fn default_manifest() -> PathBuf {
+    let mut manifest = env::current_dir().unwrap();
     manifest.push("Cargo.toml");
     manifest.canonicalize().unwrap_or(manifest)
 }
@@ -98,9 +107,13 @@ pub(super) fn get_run_types(args: &ArgMatches) -> Vec<RunType> {
 }
 
 pub(super) fn get_excluded(args: &ArgMatches) -> Vec<Regex> {
+    regexes_from_excluded(&get_list(args, "exclude-files"))
+}
+
+pub(super) fn regexes_from_excluded(strs: &[String]) -> Vec<Regex> {
     let mut files = vec![];
 
-    for temp_str in &get_list(args, "exclude-files") {
+    for temp_str in strs {
         let s = &temp_str.replace(".", r"\.").replace("*", ".*");
 
         if let Ok(re) = Regex::new(s) {
@@ -109,7 +122,6 @@ pub(super) fn get_excluded(args: &ArgMatches) -> Vec<Regex> {
             error!("Invalid regex: {}", temp_str);
         }
     }
-
     files
 }
 
@@ -120,4 +132,32 @@ pub(super) fn get_timeout(args: &ArgMatches) -> Duration {
     } else {
         Duration::from_secs(60)
     }
+}
+
+pub fn deserialize_ci_server<'de, D>(d: D) -> Result<Option<CiService>, D::Error>
+where
+    D: Deserializer<'de>,
+{
+    struct CiServerVisitor;
+
+    impl<'de> de::Visitor<'de> for CiServerVisitor {
+        type Value = Option<CiService>;
+
+        fn expecting(&self, formatter: &mut fmt::Formatter) -> fmt::Result {
+            formatter.write_str("A string containing the ci-service name")
+        }
+
+        fn visit_str<E>(self, v: &str) -> Result<Self::Value, E>
+        where
+            E: de::Error,
+        {
+            if v.is_empty() {
+                Ok(None)
+            } else {
+                Ok(Some(Ci::from_str(v).unwrap().0))
+            }
+        }
+    }
+
+    d.deserialize_any(CiServerVisitor)
 }